requirements.txt
streamlit>=1.54

pandas>=2.0,<3
numpy>=1.24

plotly>=5
matplotlib>=3.8

folium>=0.20
streamlit-folium>=0.26

xarray>=2024.1
cartopy>=0.22
branca>=0.7

pyyaml>=6.0
requests>=2.31